utils/openrouter_llm.py
import requests
import json
import os
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("OPENROUTER_API_KEY")

URL = "https://openrouter.ai/api/v1/chat/completions"


def call_llm(prompt):

    payload = {
        "model": "openai/gpt-4o-mini",
        "messages": [
            {
                "role": "user",
                "content": prompt
            }
        ]
    }

    headers = {
        "Authorization": f"Bearer {API_KEY}",
        "Content-Type": "application/json"
    }

    response = requests.post(URL, headers=headers, json=payload)

    result = response.json()

    # DEBUG: show API errors
    if "choices" not in result:
        print("OpenRouter API error:")
        print(result)
        return "LLM error occurred"

    return result["choices"][0]["message"]["content"]