requirements.txt
streamlit
langchain
langgraph
sentence-transformers
faiss-cpu
paddleocr
opencv-python
openai-whisper
sympy
python-dotenv
pydantic
numpy
pandas
google-generativeai
python-dotenv
langgraph
sympy
scikit-learn